Add the mob scene
Add the script and scene, and change the type of the scene to the new
class in the script.

To my surprise, my call to `connect` required borrowing a reference to
`callable`, even though it seems I didn't have to in *dodge* the creeps?

This completes "Designing the mob scene"[^1].

[^1]: https://docs.godotengine.org/en/stable/getting_started/first_3d_game/04.mob_scene.html