Jump and squash mobs
Set up collision layers/masks. Put `mob.tscn` in a group, so we can
listen for collisions with the group.

Allow the player to jump. When landing on a mob, squash it.

Emit a signal when the Mob is squashed. TODO: connect it to something.

Completes
https://docs.godotengine.org/en/stable/getting_started/first_3d_game/06.jump_and_squash.html

TODO: When the player strikes the mobs without jumping, they still
squash them and bounce off. Maybe this will be prevented by handling
the mobs defeating the player?