Kill the player (when hitting a mob)
Completes
https://docs.godotengine.org/en/stable/getting_started/first_3d_game/07.killing_player.html

To the player scene, add an Area2D. It has its own child, a
CollisionShape3D, which determines its details. Position it to detect
collisions from mobs (also specified using our physics layers). Listen
for `body_entered`. Due to our phsyics layers, it must be a mob. Upon
collision, emit the `hit` signal and remove the player.

In the main scene, listen for the recently added `hit` signal. Stop the
MobTimer, so we stop spawning new mobs when the player is hit.