Track the score and allow retries
Add a "UserInterface" `Control` to hold our user interface. Use anchor
presets to make it the size of the screen. It holds the score, which
gets updated when a `Mob` is squashed. It also holds a translucent
background (hidden) which shows up when the player dies, with text
offering to retry.

Specify the font (as part of the theme) for the text.

Add a controller button to "ui_accept", so we can retry directly from
the controller.

Some possible interesting follow-on steps:
- Convert String more clearly?
- Add pause, quit
- Add a background - wouldn't it be cool to make this happen on the
  surface of the water? Above sand?
- Extra points (and different sound effects?) for consecutive squashes