Change the animation speed while moving
The tutorial[^1] suggested setting the speed to 4.0, but I found that to
look quite erratic. Maybe my animation isn't quite set up like the
tutorial's, but using a different value helped. In addition, export a
variable to the editor for tweaking, which goes a lot faster than
having to recompile the code. (Though you do still have to restart the
scene.)

[^1]: https://docs.godotengine.org/en/stable/getting_started/first_3d_game/09.adding_animations.html#the-float-animation