Simplify collisions with the Player
The tutorial suggested using two separate collision detectors. But I had
trouble frequently dying when I expected to squash the mob. So reduce to
one shape that more closely matches the shape of the player. If it
strikes a mob from above, the mob is squashed; other collisions result
in player death. So far it seems to result in no unexpected player
deaths. Occasionally it seems the player may squash a mob unexpectedly,
but it's infrequent, and I haven't managed to repro. I think it relates
to the animation - if the player is rotating its front upwards, it may
squash the mob without jumping.