Ignore files in .vscode
The only file I have in here is launch.json, modified from[^1]. It is
clearly something that may be different from machine to machine/dev to
dev (e.g. path to godot). Interesting modifications:
- changed `cwd` to my project's `godot` folder, so it launches my game
- proper path to my installation of godot
- remove windows and osx, since I'm running on linux

With this setup, I'm able to use breakpoints!

It's possible other files will go in this folder, but maybe they're
specific to the machine? If not, I can adjust this later.

[^1]: https://godot-rust.github.io/book/toolchain/debugging.html