Improve player death
The player should only signal death once. (This becomes more noticeable
in a future commit when the signal results in saving the high score.)
When the player is found to be hit, make a note of it, but continue the
loop in order to check other collisions with mobs, which might result in
a final point, if the player also should've squashed another mob. After
checking collisions, kill the player as necessary. If the player dies,
exit the function, since a dead player doesn't need a new velocity,
rotation, etc.

Note that the `break` after a collision might prevent squashing multiple
mobs in the same frame.