Use better check for duplicate squashing
Previously, we simply broke out of the collisions loop when squashing a
mob. This had a couple of (at least theoretical) downsides:
- If a collision in the same frame killed the player, it could be
  skipped, depending on the order of the collisions.
- If a collision in the same frame squashed a separate mob, it would
  be skipped.

In either case, it seems likely that the next frame would catch the
missed event. But it still feels a little odd the way it was written.

Note for debugging: With a `mob` of type `Gd<Mob`, I can call `get_name`
directly to retrieve a name. Before adding to the tree, it's just "Mob",
but after adding, it has a unique name, that may look something like
"@Mob@17". (Note also that a single `Mob` may retain the name "Mob" in
the tree at a time.) For some reason I haven't determined, other method
calls (e.g. `squash`, which I wrote) require a call to `bind` (or
`bind_mut` in this case). (`instance_id` is a method on the `Gd`, so it
doesn't need a `bind` call.)

Avoid premature optimizations, but note that we create a new `HashSet`
for every physics frame. This could easily be optimized to defer until
it's needed, or reuse the same object, depending on need.