Save and load the high score
Follow the tutorial[^1] for saving and loading data. Save the high score
on death, assuming it has changed. (Determine whether it's changed by
comparing it to the current score. Note that if the player tied the high
score, we would unnecessarily re-save the same value.)

Note: I was surprised that my `i32`, after being saved and reloaded,
loaded as an `f32`. Also, I haven't figured out where this file is
saved. It'd be nice to be able to clear it; maybe the easiest way would
be in a new menu.

[^1]: https://docs.godotengine.org/en/stable/tutorials/io/saving_games.html