Add the ability to pause
Godot makes this super easy - barely an inconvenience! Set the
UserInterface's process mode to ALWAYS ("3" in the saved scene) so it
will be processed even while paused.

Update the check for reload - I haven't found a case where unwrapping
here caused a problem, but I added the check in trying to discover the
problem, and it matches the usage down below.

(The real problem was the call below. This is doubly fixed, since we
return after reloading, and now we check the `Option`.)