Play a squish sound when the player dies
I'm not super happy with the current structure, but I'm not set on the
right one.

The `ScoreContainer` has the squish sound, but that's only because it
had the callback for when the mob is squashed. The main scene knows the
player was hit; how should it access the node to make the sound? For
now, access through the `UserInterface`, which it already accesses. This
at least seems better than using the `ScoreContainer` directly.

Maybe sound effects should be global? Or owned by the main scene? Or
accessed via signals? Something to think about...