Notify when the player has squashed multiple mobs
Without landing. I'd been thinking about keeping track of it in the UI,
but I think it needs to be more obvious to the player that something is
going on. I don't necessarily even notice that my score is going up
faster, because I'm watching my character, not the scoreboard.

Some games solve this by showing the number of points gained near the
player, and I may fall back to that. Right now I'm trying to encourage
the player to get a long streak, so I'm showing the number.

Note: Without the QuadMesh behind the text, the AA on the text interacts
in bad ways with the transparent water. Apparently this is a known issue
in Godot, but there might be a nicer looking way to get around it?
Turning off AA for the text looked terrible, and using screen space AA
didn't seem to help (nor did MSAA).

TODO: Switch to showing the number of points?
TODO: Animate it. We ought to be able to use the same system we used to
animate the player and mobs.