Add a debug option to spawn mobs in a straight line
I'm trying to test consecutive bounces, and it occurred to me that it
would be easier if these mobs were all moving in a straight line at the
same speed. So add an exported variable, which I can turn on for
debugging, that does this.

Note: They actually aim at the player. I could take this further and
pass a point in the middle (instead of the player's location), but the
current usage works well enough.

TODO: Think of a better name? I'm happy with having added
`for_debugging` to make it clearer why it's there.