Animate the notification
Make the bubble, which notifies your running streak of consecutive
bounces, animate. Grow from a smaller size, and use the animation to
make it show and hide, rather than using code directly.

Initially I tried to reuse the `AnimationPlayer` that handles the
floating animation. This stopped the floating animation as soon as the
other one started, so I split up the animations into two
`AnimationPlayer`s. Not sure I did this in the cleanest way, but in the
future I'll know to simply use separate players for animations that may
happen simultaneously.

Note that I may not need to specify the animation, now that I've split
up the players.